Directory restructuring and a bunch of fixes
I've restructured the project back to the old directory format.
`go build` not working is not a good sign and godep isn't mature enough
for us to use. However, I have made sure that we can easy switch over to
godep when we want to by making sure we don't have any clashing
directories.